src/3. Retrieval Augmented Generation/pyproject.toml
[project]
name = "RAG_primitves"
version = "0.1.0"
description = "A set of RAG primitives, for independend execution"
authors = ["Andrzej Wodecki <andrzej.wodecki@example.com>"]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=0.6.3",
    "faiss-cpu>=1.12.0",
    "faker>=37.8.0",
    "langchain-chroma==0.2.3",
    "langchain-community==0.3.22",
    "langchain-core==0.3.55",
    "langchain-experimental>=0.3.0",
    "langchain-neo4j>=0.5.0",
    "langchain-openai==0.3.11",
    "langchain-text-splitters==0.3.8",
    "langchain==0.3.24",
    "libmagic>=1.0",
    "markdown>=3.9",
    "neo4j>=5.28.2",
    "pandas-stubs~=2.3.2",
    "pydantic>=2.11.4",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "python-magic>=0.4.27",
    "ragas>=0.3.5",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.4",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=5.1.1",
    "toml>=0.10.2",
    "unstructured[pdf]>=0.17.2",
    "weasyprint>=66.0",
]
